itoa input should be long long to support size_t and offset_t

+++ include/recorder.h
@@ -80,7 +80,7 @@ char* get_filename_id(const char *filename);    // map filename to an integer
 long get_file_size(const char *filename);       // return the size of a file
 int exclude_filename(const char *filename);     // if include the file in trace
 double recorder_wtime(void);                    // return the timestamp
-char* itoa(int val);                            // convert a integer to string
+char* itoa(long long val);                      // convert a integer to string
 char* ftoa(double val);                         // convert a float to string
 char* ptoa(const void* ptr);                    // convert a pointer to string
 char** assemble_args_list(int arg_count, ...);
